# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Indo-German Mobility Corridor Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vacancies:
    get:
      operationId: listVacancies
      tags: [vacancies]
      summary: List vacancies
      parameters:
        - name: occupation
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Vacancies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vacancy"
    post:
      operationId: createVacancy
      tags: [vacancies]
      summary: Create vacancy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VacancyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vacancy"
  /vacancies/{id}:
    get:
      operationId: getVacancy
      tags: [vacancies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vacancy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vacancy"
    patch:
      operationId: updateVacancy
      tags: [vacancies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VacancyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vacancy"
  /vacancies/{id}/matches:
    get:
      operationId: getVacancyMatches
      tags: [vacancies]
      summary: Get matched candidates for a vacancy (ranked by fit score)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Matched candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CandidateMatch"

  # ── APPLICATIONS ─────────────────────────────────────────────────────────────
...
  /courses:
    get:
      operationId: listCourses
      tags: [training]
      summary: List language/training courses
      responses:
        "200":
          description: Courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [training]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /batches:
    get:
      operationId: listBatches
      tags: [training]
      summary: List batches
      responses:
        "200":
          description: Batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Batch"
    post:
      operationId: createBatch
      tags: [training]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
  /batches/{id}/enroll:
    post:
      operationId: enrollInBatch
      tags: [training]
      summary: Enroll candidate in a batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentInput"
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
...
  /recognition:
    get:
      operationId: listRecognitionCases
      tags: [recognition]
      summary: List recognition cases
      responses:
        "200":
          description: Recognition cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecognitionCase"
    post:
      operationId: createRecognitionCase
      tags: [recognition]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecognitionCaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecognitionCase"
  /recognition/{id}:
    get:
      operationId: getRecognitionCase
      tags: [recognition]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recognition case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecognitionCase"
    patch:
      operationId: updateRecognitionCase
      tags: [recognition]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecognitionCaseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecognitionCase"

  # ── VISA ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CandidateMatch:
      type: object
      required: [candidateId, name, fitScore, stage, readinessScore]
      properties:
        candidateId:
          type: integer
        name:
          type: string
        occupation:
          type: ["string", "null"]
        germanLanguageLevel:
          type: ["string", "null"]
        fitScore:
          type: integer
        stage:
          type: string
        readinessScore:
          type: integer
        district:
          type: ["string", "null"]
        experienceYears:
          type: ["integer", "null"]

    # ── EMPLOYER ──
...
    Vacancy:
      type: object
      required: [id, employerId, title, occupation, location, status, createdAt]
      properties:
        id:
          type: integer
        employerId:
          type: integer
        employerName:
          type: ["string", "null"]
        title:
          type: string
        occupation:
          type: string
        location:
          type: string
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        languageLevel:
          type: ["string", "null"]
        shiftType:
          type: ["string", "null"]
        experienceYears:
          type: ["integer", "null"]
        recognitionRequired:
          type: boolean
        visaSupport:
          type: boolean
        accommodationSupport:
          type: boolean
        joiningDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, filled, paused, closed]
        applicantsCount:
          type: integer
        createdAt:
          type: string
    VacancyInput:
      type: object
      required: [title, occupation, location]
      properties:
        title:
          type: string
        occupation:
          type: string
        location:
          type: string
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        languageLevel:
          type: string
        shiftType:
          type: string
        experienceYears:
          type: integer
        recognitionRequired:
          type: boolean
        visaSupport:
          type: boolean
        accommodationSupport:
          type: boolean
        joiningDate:
          type: string
    VacancyUpdate:
      type: object
      properties:
        title:
          type: string
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        status:
          type: string
        languageLevel:
          type: string

    # ── APPLICATION ──
...
    Course:
      type: object
      required: [id, name, languageLevel, providerId, providerName, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        languageLevel:
          type: string
          enum: [A1, A2, B1, B2, C1]
        providerId:
          type: integer
        providerName:
          type: string
        durationWeeks:
          type: ["integer", "null"]
        feeAmount:
          type: ["integer", "null"]
        mode:
          type: string
          enum: [online, offline, hybrid]
        city:
          type: ["string", "null"]
        enrolledCount:
          type: integer
        createdAt:
          type: string
    CourseInput:
      type: object
      required: [name, languageLevel, mode]
      properties:
        name:
          type: string
        languageLevel:
          type: string
        durationWeeks:
          type: integer
        feeAmount:
          type: integer
        mode:
          type: string
        city:
          type: string
    Batch:
      type: object
      required: [id, courseId, courseName, startDate, status, capacity, enrolledCount]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        courseName:
          type: string
        trainerName:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [upcoming, active, completed, cancelled]
        capacity:
          type: integer
        enrolledCount:
          type: integer
        attendancePercent:
          type: ["integer", "null"]
    BatchInput:
      type: object
      required: [courseId, startDate, capacity]
      properties:
        courseId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        capacity:
          type: integer
        trainerName:
          type: string
    Enrollment:
      type: object
      required: [id, candidateId, batchId, status, enrolledAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        candidateName:
          type: ["string", "null"]
        batchId:
          type: integer
        batchName:
          type: ["string", "null"]
        languageLevel:
          type: ["string", "null"]
        status:
          type: string
          enum: [enrolled, attending, passed, failed, dropped]
        attendancePercent:
          type: ["integer", "null"]
        assessmentScore:
          type: ["integer", "null"]
        certificateIssued:
          type: boolean
        enrolledAt:
          type: string
    EnrollmentInput:
      type: object
      required: [candidateId]
      properties:
        candidateId:
          type: integer
...
    RecognitionCase:
      type: object
      required: [id, candidateId, indianQualification, germanProfession, status, createdAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        candidateName:
          type: ["string", "null"]
        indianQualification:
          type: string
        germanProfession:
          type: string
        regulatedStatus:
          type: string
          enum: [regulated, non_regulated, unknown]
        recognitionAuthority:
          type: ["string", "null"]
        status:
          type: string
          enum: [not_required, eligibility_check_pending, documents_pending, application_ready, submitted, authority_review, partial_recognition, full_recognition, rejected, bridge_course_required]
        documentsRequired:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    RecognitionCaseInput:
      type: object
      required: [indianQualification, germanProfession]
      properties:
        indianQualification:
          type: string
        germanProfession:
          type: string
        recognitionAuthority:
          type: string
    RecognitionCaseUpdate:
      type: object
      properties:
        status:
          type: string
        recognitionAuthority:
          type: string
        notes:
          type: string
        documentsRequired:
          type: array
          items:
            type: string

    # ── VISA ──
...