# lib/api-zod/src/generated/types/recognitionCaseUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Indo-German Mobility Corridor Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RecognitionCaseUpdate {
  status?: string;
  recognitionAuthority?: string;
  notes?: string;
  documentsRequired?: string[];
}