# lib/api-zod/src/generated/types/vacancy.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Indo-German Mobility Corridor Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { VacancyStatus } from './vacancyStatus';

export interface Vacancy {
  id: number;
  employerId: number;
  /** @nullable */
  employerName?: string | null;
  title: string;
  occupation: string;
  location: string;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  /** @nullable */
  languageLevel?: string | null;
  /** @nullable */
  shiftType?: string | null;
  /** @nullable */
  experienceYears?: number | null;
  recognitionRequired?: boolean;
  visaSupport?: boolean;
  accommodationSupport?: boolean;
  /** @nullable */
  joiningDate?: string | null;
  status: VacancyStatus;
  applicantsCount?: number;
  createdAt: string;
}